+++ pkg-r/tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(shinyrealtime)
+
+test_check("shinyrealtime")